[MaL] Moved sparse.h to MathLib/LinAlg/Sparse/.
Corrected also the include paths for files in MathLib folder and SimpleTests/MatrixTests/ folder.
